+++ src/apis/member/userop.py
@@ -207,8 +207,7 @@ async def post_member_resign(request: Request, response: Response, authorization
     discordid = au["discordid"]
     name = convertQuotation(au["name"])
 
-    stoken = authorization.split(" ")[1]
-    if stoken.startswith("e"):
+    if not (await isSecureAuth(dhrid, authorization, request)):
         response.status_code = 403
         return {"error": ml.tr(request, "access_sensitive_data", force_lang = au["language"])}