+++ src/db.py
@@ -56,6 +56,7 @@ cur.execute(f"CREATE TABLE IF NOT EXISTS session (token CHAR(36), uid INT, times
 cur.execute(f"CREATE TABLE IF NOT EXISTS ratelimit (ip TEXT, endpoint TEXT, first_request_timestamp BIGINT, request_count INT)")
 cur.execute(f"CREATE TABLE IF NOT EXISTS auth_ticket (token CHAR(36), uid BIGINT UNSIGNED, expire BIGINT)")
 cur.execute(f"CREATE TABLE IF NOT EXISTS application_token (app_name TEXT, token CHAR(36), uid BIGINT UNSIGNED, timestamp BIGINT, last_used_timestamp BIGINT)")
+cur.execute(f"CREATE TABLE IF NOT EXISTS email_confirmation (uid INT, secret TEXT, operation TEXT, expire BIGINT)")
 cur.execute(f"CREATE TABLE IF NOT EXISTS auditlog (uid INT, operation TEXT, timestamp BIGINT) DATA DIRECTORY = '{config.mysql_ext}'")
 cur.execute(f"CREATE TABLE IF NOT EXISTS settings (uid BIGINT UNSIGNED, skey TEXT, sval TEXT)")
 
@@ -122,6 +123,8 @@ indexes = ["CREATE INDEX user_uid ON user (uid)",
 "CREATE INDEX auth_ticket_token ON auth_ticket (token)",
 "CREATE INDEX application_token_token ON application_token (token)",
 "CREATE INDEX ratelimit_ip ON ratelimit (ip)",
+"CREATE INDEX email_confirmation_uid ON email_confirmation (uid)",
+"CREATE INDEX email_confirmation_secret ON email_confirmation (secret)",
 "CREATE INDEX auditlog_userid ON auditlog (userid)",
 "CREATE INDEX settings_uid ON settings (uid)"]