+++ src/functions/main.py
@@ -5,6 +5,7 @@ from functions.arequests import *
 from functions.dataop import *
 from functions.general import *
 from functions.iptype import *
+from functions.mail import *
 from functions.notification import *
 from functions.ranking import *
 from functions.security import *