+++ src/functions/totp.py
@@ -1,3 +1,6 @@
+# Copyright (C) 2023 CharlesWithC All rights reserved.
+# Author: @CharlesWithC
+
 import base64
 import hashlib
 import hmac